refactor(layout): move root metadata fetch to root layout
ref: #203 #206
